chore(substrate): consolidate test scaffolding
