+++ packages/agent/test/core/budget.test.ts
@@ -8,6 +8,7 @@ import {
   createBudgetState,
   effectiveBudgetThresholds,
   publishBudgetTelemetry,
+  recordTokenUsage,
 } from "../../src/core/budget";
 
 /** The run whose budget is being reported; the reporter never mints one. */
@@ -259,5 +260,21 @@ describe("defaults are written once — narration agrees with enforcement", () =
     expect(desc).toContain(
       `${Math.round(BUDGET_DEFAULTS.maxToolRuntimeMs / 1000)}s tool wall time`,
     );
+  });
+});
+
+describe("BudgetState token tracking", () => {
+  it("starts with zero token counts", () => {
+    const state = createBudgetState();
+    expect(state.totalInputTokens).toBe(0);
+    expect(state.totalOutputTokens).toBe(0);
+  });
+
+  it("recordTokenUsage accumulates tokens", () => {
+    let state = createBudgetState();
+    state = recordTokenUsage(state, 100, 50);
+    state = recordTokenUsage(state, 200, 100);
+    expect(state.totalInputTokens).toBe(300);
+    expect(state.totalOutputTokens).toBe(150);
   });
 });
